添加图标

+++ src/main/java/com/pengfu/view/page/AdminListPage.java
@@ -2,8 +2,8 @@ package com.pengfu.view.page;
 
 import java.awt.BorderLayout;
 import java.awt.Dimension;
+import java.awt.FlowLayout;
 
-import javax.swing.JButton;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
@@ -20,6 +20,7 @@ import com.pengfu.util.ConstantConfig;
 import com.pengfu.util.SpringContextUtils;
 import com.pengfu.util.TableBuilder;
 import com.pengfu.view.PopupFrame;
+import com.pengfu.view.component.AppButton;
 
 @Component
 @Lazy
@@ -50,15 +51,15 @@ public class AdminListPage extends BasePage {
 		northPane.setBackground(ConstantConfig.PAGE_COLOR);
 		northPane.setPreferredSize(new Dimension(0, 64));	
 		contxtPane.add(northPane, BorderLayout.NORTH);
-		
+		northPane.setLayout(new FlowLayout(FlowLayout.RIGHT, 16, 5));
 		// 操作按钮
-		JButton addBtn = new JButton("添加");
+		AppButton addBtn = new AppButton("添加", ConstantConfig.ADD_IMG);
 		northPane.add(addBtn);
-		JButton setBtn = new JButton("修改");
+		AppButton setBtn = new AppButton("修改", ConstantConfig.SET_IMG);
 		northPane.add(setBtn);
-		JButton deleteBtn = new JButton("删除");
+		AppButton deleteBtn = new AppButton("删除", ConstantConfig.DELETE_IMG);
 		northPane.add(deleteBtn);
-		JButton updateBtn = new JButton("刷新");
+		AppButton updateBtn = new AppButton("刷新", ConstantConfig.UPDATE_IMG);
 		northPane.add(updateBtn);
 		
 		// 楼宇信息列表
@@ -93,15 +94,18 @@ public class AdminListPage extends BasePage {
 			popupFrame.setVisible(true);
 		});
 		deleteBtn.addActionListener(e -> {
-			int row = table.getSelectedRow();
-			if(row == -1) {
-				JOptionPane.showMessageDialog(null, "未选择目标");
-				return;
+			if(JOptionPane.showConfirmDialog(null, "确定删除此管理员?", "删除", JOptionPane.YES_NO_OPTION) 
+					== JOptionPane.YES_OPTION) {
+				int row = table.getSelectedRow();
+				if(row == -1) {
+					JOptionPane.showMessageDialog(null, "未选择目标");
+					return;
+				}
+				if(adminService.delete((String) model.getValueAt(row, 1)) > 0) {
+					updateTable();
+					JOptionPane.showMessageDialog(null, "删除成功");
+				}	
 			}
-			if(adminService.delete((String) model.getValueAt(row, 1)) > 0) {
-				updateTable();
-				JOptionPane.showMessageDialog(null, "删除成功");
-			}	
 		});
 		// 刷新
 		updateBtn.addActionListener(e -> updateTable());
