+++ src/main/java/com/pengfu/view/page/DormListPage.java
@@ -3,10 +3,10 @@ package com.pengfu.view.page;
 import java.awt.BorderLayout;
 import java.awt.Color;
 import java.awt.Dimension;
+import java.awt.FlowLayout;
 
 import javax.swing.BorderFactory;
 import javax.swing.Box;
-import javax.swing.JButton;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
@@ -23,6 +23,7 @@ import com.pengfu.util.ConstantConfig;
 import com.pengfu.util.SpringContextUtils;
 import com.pengfu.util.TableBuilder;
 import com.pengfu.view.PopupFrame;
+import com.pengfu.view.component.AppButton;
 import com.pengfu.view.component.TitleComboBox;
 
 @Component
@@ -31,7 +32,7 @@ public class DormListPage extends BasePage {
 
 	private static final long serialVersionUID = 1L;
 	
-	private TitleComboBox buildingComboBox;
+	private TitleComboBox bidComboBox;
 	
 	private DormTableModel model = new DormTableModel();
 	private JTable table;
@@ -54,9 +55,12 @@ public class DormListPage extends BasePage {
 		topPane.setBorder(BorderFactory.createLineBorder(new Color(65, 113, 156), 1));
 		add(topPane, 0);
 		// 楼宇
-		buildingComboBox = new TitleComboBox("楼宇");
-		buildingComboBox.setModel(Role.getAdmin().getBids());
-		topPane.add(buildingComboBox);
+		bidComboBox = new TitleComboBox("楼宇");
+		bidComboBox.setModel(Role.getAdmin().getBids());
+		topPane.add(bidComboBox);
+		// 搜索按钮
+		AppButton searchBtn = new AppButton("搜索", ConstantConfig.SEARCH_IMG);
+		topPane.add(searchBtn);
 		
 		// 分隔
 		add(Box.createVerticalStrut(16), 1);
@@ -69,26 +73,27 @@ public class DormListPage extends BasePage {
 		northPane.setBackground(ConstantConfig.PAGE_COLOR);
 		northPane.setPreferredSize(new Dimension(0, 64));	
 		contxtPane.add(northPane, BorderLayout.NORTH);
-		
+		northPane.setLayout(new FlowLayout(FlowLayout.RIGHT, 16, 5));
 		// 操作按钮
-		JButton addBtn = new JButton("添加");
+		AppButton addBtn = new AppButton("添加", ConstantConfig.ADD_IMG);
 		northPane.add(addBtn);
-//		JButton setBtn = new JButton("修改");
+//		AppButton setBtn = new AppButton("修改", ConstantConfig.SET_IMG);
 //		northPane.add(setBtn);
-		JButton deleteBtn = new JButton("删除");
+		AppButton deleteBtn = new AppButton("删除", ConstantConfig.DELETE_IMG);
 		northPane.add(deleteBtn);
-		JButton updateBtn = new JButton("刷新");
+		AppButton updateBtn = new AppButton("刷新", ConstantConfig.UPDATE_IMG);
 		northPane.add(updateBtn);
 		
 		// 宿舍信息列表
-		model.setDorms(dormService.getDormitoryByBid(buildingComboBox.getText()));
+		model.setDorms(dormService.getDormitoryByBid(bidComboBox.getText()));
 		table = TableBuilder.getTableBuilder().build(model);
 		JScrollPane tablePane = new JScrollPane(table);
 		tablePane.getViewport().setBackground(ConstantConfig.PAGE_COLOR);
 		contxtPane.add(tablePane, BorderLayout.CENTER);
 		
 		// 监听器 
-		buildingComboBox.addActionListener(e -> {
+		// 搜索
+		searchBtn.addActionListener(e -> {
 			updateTable();
 		});
 		addBtn.addActionListener((e) ->{
@@ -110,15 +115,18 @@ public class DormListPage extends BasePage {
 //			popupFrame.setVisible(true);
 //		});
 		deleteBtn.addActionListener(e -> {
-			int row = table.getSelectedRow();
-			if(row == -1) {
-				JOptionPane.showMessageDialog(null, "未选择目标");
-				return;
+			if(JOptionPane.showConfirmDialog(null, "确定删除此宿舍?", "删除", JOptionPane.YES_NO_OPTION) 
+					== JOptionPane.YES_OPTION) {
+				int row = table.getSelectedRow();
+				if(row == -1) {
+					JOptionPane.showMessageDialog(null, "未选择目标");
+					return;
+				}
+				if(dormService.delete(model.get(row)) > 0) {
+					updateTable();
+					JOptionPane.showMessageDialog(null, "删除成功");
+				}	
 			}
-			if(dormService.delete(model.get(row)) > 0) {
-				updateTable();
-				JOptionPane.showMessageDialog(null, "删除成功");
-			}	
 		});
 		// 刷新
 		updateBtn.addActionListener(e -> updateTable());
@@ -127,7 +135,7 @@ public class DormListPage extends BasePage {
 	
 	/** 更新表格数据 */
 	public void updateTable() {
-		model.setDorms(dormService.getDormitoryByBid(buildingComboBox.getText()));
+		model.setDorms(dormService.getDormitoryByBid(bidComboBox.getText()));
 		table.updateUI();
 	}
 }
